+++ group-project/src/game/PhaserGame.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref, onMounted, onUnmounted } from 'vue';
-import { EventBus } from './scenes/EventBus';
-import StartGame from './main'; 
+import { EventBus } from './scenes/EventBus.js';
+import StartGame from '../main.js';
 
 
 const scene = ref()
@@ -19,20 +19,20 @@ const questions = [
     { id: 5, text: '9 * 8 = 72', correctAnswer: 'Yes' },
     { id: 6, text: '10 * 10 = 100', correctAnswer: 'Yes' },
     { id: 7, text: '12 * 2 = 22', correctAnswer: 'No' },
-   
+
 ];
 
 const emit = defineEmits(['current-active-scene'])
 
 onMounted(() => {
-   
+
     game.value = StartGame('game-container')
     EventBus.on('current-scene-ready', (currentScene) => {
         emit('current-active-scene', currentScene)
         scene.value = currentScene
  });
 
-  
+
     EventBus.on('chest-collided', ({ id }) => {
         const question = questions.find(q => q.id === id)
         if (question) {
@@ -40,10 +40,10 @@ onMounted(() => {
             showQuestion.value = true
             currentChestId.value = id
         }
-        
+
     })
     EventBus.on('game-over', () => {
-    score.value = 0; 
+    score.value = 0;
 })
 })
 
@@ -64,7 +64,7 @@ const submitAnswer = (answer) => {
     } else {
         console.log("Incorrect. Game over.")
         EventBus.emit('answer-submitted', { isCorrect: false, chestId: currentChestId.value })
-       
+
     }
     showQuestion.value = false
 
@@ -74,14 +74,22 @@ defineExpose({ scene, game })
 </script>
 
 <template>
-    <div id="game-container"></div>
-    <div class="score-display">Score: {{ score }}</div>
+  <div id="game-container"></div>
+  <div class="score-display">Score: {{ score }}</div>
 
-    <div v-if="showQuestion" class="modal">
-        <p>{{ questionText }}</p>
-        <button @click="submitAnswer('Yes')">Yes</button>
-        <button @click="submitAnswer('No')">No</button>
-    </div>
+  <div v-if="showQuestion" class="fixed inset-0 flex justify-center items-center">
+      <div class="modal bg-gray-800 text-gray-200 p-5 rounded-md font-sans z-50">
+          <p>{{ questionText }}</p>
+          <div class="flex space-x-4 mt-4">
+              <button @click="submitAnswer('Yes')" class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">
+                  Yes
+              </button>
+              <button @click="submitAnswer('No')" class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded">
+                  No
+              </button>
+          </div>
+      </div>
+  </div>
 </template>
 
 <style>
@@ -89,7 +97,7 @@ defineExpose({ scene, game })
     position: fixed;
   left: 50%;
   top: 50%;
-  
+
   padding: 20px;
   background: #222222;
   border-radius: 5px;
@@ -98,9 +106,3 @@ defineExpose({ scene, game })
   z-index: 100;
 }
 </style>
-
-
-
-
-
-
