+++ group-project/src/game/scenes/MainMenu.js
@@ -15,10 +15,10 @@ export class MainMenu extends Scene {
     create() {
         // Add a background image to the scene
         this.add.image(256, 256, 'background');
-   
+
         // Add a game logo to the scene
-        this.logo = this.add.image(256, 150, 'logo').setDepth(100);
-   
+        this.logo = this.add.image(256, 150, 'platform').setDepth(100);
+
         // Add main menu text
         this.add.text(256, 230, 'Main Menu', {
             fontFamily: 'Arial Black',
@@ -28,8 +28,8 @@ export class MainMenu extends Scene {
             strokeThickness: 6,
             align: 'center'
         }).setDepth(100).setOrigin(0.5);
-   
-       
+
+
         const playGameButton = this.add.text(256, 300, 'Play Game', {
             fontFamily: 'Arial',
             fontSize: 32,
@@ -45,12 +45,12 @@ export class MainMenu extends Scene {
             // If your game doesn't need a preloader, you can change this to directly start the 'Game' scene
             this.scene.start('Preloader');
         });
-   
+
         // Notify any system listening (like your Vue app) that the current scene is ready
         // This is useful if your Vue app needs to adjust UI based on the active Phaser scene
         EventBus.emit('current-scene-ready', this);
     }
-   
+
 
 
     changeScene() {
@@ -88,5 +88,3 @@ export class MainMenu extends Scene {
         }
     }
 }
-
-
