+++ group-project/src/game/scenes/Preloader.js
@@ -11,7 +11,7 @@ export class Preloader extends Scene
     init ()
     {
         //  We loaded this image in our Boot Scene, so we can display it here
-        this.add.image(512, 384, 'background')
+        this.add.image(512, 384, 'treasuremath')
 
         //  A simple progress bar. This is the outline of the bar.
         this.add.rectangle(512, 384, 468, 32).setStrokeStyle(1, 0xffffff)
@@ -31,22 +31,22 @@ export class Preloader extends Scene
     preload ()
     {
         //  Load the assets for the game - Replace with your own assets
-        this.load.setPath('assets')
+        // this.load.setPath('public')
 
-        this.load.image('logo', 'logo.png')
+        this.load.image('logo', 'LVL.png')
         this.load.atlas('archer', 'archer.png', 'archer_atlas.json' )
         this.load.animation('archer_anim', 'archer_anim.json')
         this.load.image('tiles', 'dungeon.png')
         this.load.tilemapTiledJSON('map', 'map.json')
-      
- this.load.image('floorTiles', 'Floor.png') 
- this.load.image('wallTiles', 'Walls.png') 
- this.load.image('doorTiles', 'DoubleDoor.png') 
+
+ this.load.image('floorTiles', 'Floor.png')
+ this.load.image('wallTiles', 'Walls.png')
+ this.load.image('doorTiles', 'DoubleDoor.png')
 this.load.spritesheet('chestTiles', 'chest.png', { frameWidth: 16, frameHeight: 16 });
  this.load.atlas('enemies','enemies.png', 'enemies_atlas.json')
  this.load.animation('enemies_atlas.json', 'enemies_anim.json')
 
- 
+
  }
 
     create ()
